Enum authzen_core::ActionError

pub enum ActionError<E1, E2> {
    Authz(E1),
    Storage(E2),
}

Represents the possible sources of error when performing an action which requires authorization.

Variants

Authz(E1)

Wraps an error returned from a DecisionMaker when the subject is either not authorized to perform an action or some other issue occurs while communicating with the DecisionMaker.

Storage(E2)

Wraps an error returned from a StorageClient when the subject has been authorized to perform an action but there an error occurs while actually performing the error. Examples of this include network errors while communicating with an api or database, unique constraint violations raised by a database, etc.

Implementations

impl<E1, E2> ActionError<E1, E2>

pub fn as_authz(&self) -> Option<&E1>

Returns Some(#return_type) if this value is of type variant_ident. Returns None otherwise

pub fn as_storage(&self) -> Option<&E2>

Returns Some(#return_type) if this value is of type variant_ident. Returns None otherwise

impl<E1, E2> ActionError<E1, E2>

pub fn as_authz_mut(&mut self) -> Option<&mut E1>

Returns Some(#return_type) if this value is of type variant_ident. Returns None otherwise

pub fn as_storage_mut(&mut self) -> Option<&mut E2>

Returns Some(#return_type) if this value is of type variant_ident. Returns None otherwise

impl<E1, E2> ActionError<E1, E2>

pub fn is_authz(&self) -> bool

Returns true if this value is of type variant_ident. Returns false otherwise

pub fn is_storage(&self) -> bool

Returns true if this value is of type variant_ident. Returns false otherwise

impl<E1, E2> ActionError<E1, E2>

pub fn unwrap_authz(self) -> E1

Unwraps this value to the variant_ident variant

Panics if this value is of any other type

pub fn unwrap_storage(self) -> E2

Unwraps this value to the variant_ident variant

Panics if this value is of any other type

impl<E1, E2> ActionError<E1, E2>

pub fn authz(err: E1) -> Self

pub fn storage(err: E2) -> Self